+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/WebFictionScreen.kt
@@ -154,70 +154,6 @@ fun WebFictionScreen(
 // TOP BAR
 // =============================================================================
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun WebFictionTopBar(
-    onBackClick: () -> Unit,
-    onRefresh: () -> Unit,
-    onAddByUrl: () -> Unit,
-    isRefreshing: Boolean
-) {
-    TopAppBar(
-        title = {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Icon(
-                    imageVector = Icons.Default.Article,
-                    contentDescription = "Web Fiction",
-                    tint = MediaColors.MediaTypes.Document,
-                    modifier = Modifier.size(24.dp)
-                )
-                Spacer(modifier = Modifier.width(MediaSpacing.SM))
-                Text(
-                    text = "Web Fiction",
-                    style = MediaTypography.TitleMedium,
-                    color = MediaColors.TextPrimary
-                )
-            }
-        },
-        navigationIcon = {
-            IconButton(onClick = onBackClick) {
-                Icon(
-                    imageVector = Icons.Default.ArrowBack,
-                    contentDescription = "Back",
-                    tint = MediaColors.TextPrimary
-                )
-            }
-        },
-        actions = {
-            if (isRefreshing) {
-                CircularProgressIndicator(
-                    modifier = Modifier.size(24.dp),
-                    color = MediaColors.AccentPrimary,
-                    strokeWidth = 2.dp
-                )
-                Spacer(modifier = Modifier.width(MediaSpacing.MD))
-            } else {
-                IconButton(onClick = onRefresh) {
-                    Icon(
-                        imageVector = Icons.Default.Refresh,
-                        contentDescription = "Check for updates",
-                        tint = MediaColors.TextSecondary
-                    )
-                }
-            }
-            IconButton(onClick = onAddByUrl) {
-                Icon(
-                    imageVector = Icons.Default.AddLink,
-                    contentDescription = "Add by URL",
-                    tint = MediaColors.TextSecondary
-                )
-            }
-        },
-        colors = TopAppBarDefaults.topAppBarColors(
-            containerColor = MediaColors.BackgroundElevated
-        )
-    )
-}
 
 // =============================================================================
 // FOLLOWING PAGE
@@ -251,226 +187,6 @@ private fun FollowingStoriesPage(
     }
 }
 
-@Composable
-private fun WebFictionStoryCard(
-    story: WebFictionStory,
-    onClick: () -> Unit
-) {
-    Surface(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(MediaCorners.MD))
-            .clickable(onClick = onClick),
-        color = MediaColors.BackgroundElevated
-    ) {
-        Row(
-            modifier = Modifier.padding(MediaSpacing.MD)
-        ) {
-            // Cover
-            Surface(
-                modifier = Modifier
-                    .size(width = 80.dp, height = 120.dp)
-                    .clip(RoundedCornerShape(MediaCorners.SM)),
-                color = MediaColors.BackgroundSurface
-            ) {
-                if (story.coverUrl != null) {
-                    AsyncImage(
-                    
-                        model = story.coverUrl,
-                        contentDescription = "Media image",
-                        contentScale = ContentScale.Crop,
-                        modifier = Modifier.fillMaxSize()
-                    )
-                } else {
-                    Box(
-                        contentAlignment = Alignment.Center,
-                        modifier = Modifier.fillMaxSize()
-                    ) {
-                        Icon(
-                            imageVector = Icons.Default.MenuBook,
-                            contentDescription = "Media image",
-                            tint = story.source.color.copy(alpha = 0.5f),
-                            modifier = Modifier.size(32.dp)
-                        )
-                    }
-                }
-                
-                // Progress overlay
-                if (story.readProgress > 0) {
-                    Box(
-                        modifier = Modifier.fillMaxSize(),
-                        contentAlignment = Alignment.BottomCenter
-                    ) {
-                        LinearProgressIndicator(
-                            progress = { story.readProgress },
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .height(4.dp),
-                            color = MediaColors.AccentPrimary,
-                            trackColor = MediaColors.ProgressBackground
-                        )
-                    }
-                }
-            }
-            
-            Spacer(modifier = Modifier.width(MediaSpacing.MD))
-            
-            Column(modifier = Modifier.weight(1f)) {
-                // Source badge
-                Surface(
-                    shape = RoundedCornerShape(MediaCorners.XS),
-                    color = story.source.color.copy(alpha = 0.2f)
-                ) {
-                    Text(
-                        text = story.source.displayName,
-                        style = MediaTypography.LabelSmall,
-                        color = story.source.color,
-                        modifier = Modifier.padding(horizontal = MediaSpacing.SM, vertical = 2.dp)
-                    )
-                }
-                
-                Spacer(modifier = Modifier.height(MediaSpacing.XS))
-                
-                // Title
-                Text(
-                    text = story.title,
-                    style = MediaTypography.BodyLarge,
-                    color = MediaColors.TextPrimary,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis,
-                    fontWeight = FontWeight.Medium
-                )
-                
-                // Author
-                Text(
-                    text = "by ${story.author}",
-                    style = MediaTypography.LabelSmall,
-                    color = MediaColors.TextSecondary,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-                
-                Spacer(modifier = Modifier.height(MediaSpacing.SM))
-                
-                // Stats row
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(MediaSpacing.MD),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    // Chapter count
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Icon(
-                            imageVector = Icons.Default.MenuBook,
-                            contentDescription = "Media image",
-                            tint = MediaColors.TextTertiary,
-                            modifier = Modifier.size(14.dp)
-                        )
-                        Spacer(modifier = Modifier.width(2.dp))
-                        Text(
-                            text = "${story.chapterCount} ch",
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextTertiary
-                        )
-                    }
-                    
-                    // Word count
-                    if (story.wordCount != null) {
-                        Row(verticalAlignment = Alignment.CenterVertically) {
-                            Icon(
-                                imageVector = Icons.Default.TextFields,
-                                contentDescription = "Media image",
-                                tint = MediaColors.TextTertiary,
-                                modifier = Modifier.size(14.dp)
-                            )
-                            Spacer(modifier = Modifier.width(2.dp))
-                            Text(
-                                text = formatWordCount(story.wordCount),
-                                style = MediaTypography.LabelSmall,
-                                color = MediaColors.TextTertiary
-                            )
-                        }
-                    }
-                    
-                    // Status
-                    story.status?.let { status ->
-                        StatusBadge(status = status)
-                    }
-                }
-                
-                Spacer(modifier = Modifier.height(MediaSpacing.SM))
-                
-                // Tags (first few)
-                if (story.tags.isNotEmpty()) {
-                    Row(
-                        horizontalArrangement = Arrangement.spacedBy(MediaSpacing.XS),
-                        modifier = Modifier.horizontalScroll(rememberScrollState())
-                    ) {
-                        story.tags.take(4).forEach { tag ->
-                            Surface(
-                                shape = RoundedCornerShape(MediaCorners.XS),
-                                color = MediaColors.BackgroundSurface
-                            ) {
-                                Text(
-                                    text = tag,
-                                    style = MediaTypography.LabelSmall,
-                                    color = MediaColors.TextTertiary,
-                                    modifier = Modifier.padding(horizontal = MediaSpacing.XS, vertical = 2.dp)
-                                )
-                            }
-                        }
-                    }
-                }
-                
-                Spacer(modifier = Modifier.height(MediaSpacing.SM))
-                
-                // Update info
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    if (story.unreadChapters > 0) {
-                        Surface(
-                            shape = RoundedCornerShape(MediaCorners.XS),
-                            color = MediaColors.AccentPrimary.copy(alpha = 0.2f)
-                        ) {
-                            Row(
-                                verticalAlignment = Alignment.CenterVertically,
-                                modifier = Modifier.padding(horizontal = MediaSpacing.SM, vertical = MediaSpacing.XS)
-                            ) {
-                                Icon(
-                                    imageVector = Icons.Default.NewReleases,
-                                    contentDescription = "Media image",
-                                    tint = MediaColors.AccentPrimary,
-                                    modifier = Modifier.size(14.dp)
-                                )
-                                Spacer(modifier = Modifier.width(MediaSpacing.XS))
-                                Text(
-                                    text = "${story.unreadChapters} new",
-                                    style = MediaTypography.LabelSmall,
-                                    color = MediaColors.AccentPrimary,
-                                    fontWeight = FontWeight.SemiBold
-                                )
-                            }
-                        }
-                    } else {
-                        Text(
-                            text = "Updated ${story.lastUpdated}",
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextTertiary
-                        )
-                    }
-                    
-                    Text(
-                        text = "${story.readChapters}/${story.chapterCount}",
-                        style = MediaTypography.LabelSmall,
-                        color = MediaColors.TextSecondary
-                    )
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun StatusBadge(status: StoryStatus) {
@@ -536,108 +252,6 @@ private fun RecentUpdatesPage(
     }
 }
 
-@Composable
-private fun UpdateItem(
-    update: WebFictionUpdate,
-    onClick: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick)
-            .padding(horizontal = MediaSpacing.MD, vertical = MediaSpacing.SM),
-        verticalAlignment = Alignment.Top
-    ) {
-        // Story cover
-        Surface(
-            modifier = Modifier
-                .size(48.dp)
-                .clip(RoundedCornerShape(MediaCorners.XS)),
-            color = MediaColors.BackgroundSurface
-        ) {
-            if (update.storyCoverUrl != null) {
-                AsyncImage(
-                    
-                    model = update.storyCoverUrl,
-                    contentDescription = "Media image",
-                    contentScale = ContentScale.Crop,
-                    modifier = Modifier.fillMaxSize()
-                )
-            } else {
-                Box(contentAlignment = Alignment.Center) {
-                    Icon(
-                        imageVector = Icons.Default.MenuBook,
-                        contentDescription = "Media image",
-                        tint = update.source.color.copy(alpha = 0.5f),
-                        modifier = Modifier.size(24.dp)
-                    )
-                }
-            }
-        }
-        
-        Spacer(modifier = Modifier.width(MediaSpacing.MD))
-        
-        Column(modifier = Modifier.weight(1f)) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
-            ) {
-                Text(
-                    text = update.storyTitle,
-                    style = MediaTypography.LabelSmall,
-                    color = update.source.color,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.weight(1f, fill = false)
-                )
-                
-                if (!update.isRead) {
-                    Box(
-                        modifier = Modifier
-                            .size(8.dp)
-                            .background(MediaColors.AccentPrimary, CircleShape)
-                    )
-                }
-            }
-            
-            Text(
-                text = update.chapter.title,
-                style = MediaTypography.BodyMedium,
-                color = MediaColors.TextPrimary,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                fontWeight = if (!update.isRead) FontWeight.SemiBold else FontWeight.Normal
-            )
-            
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = update.timestamp,
-                    style = MediaTypography.LabelSmall,
-                    color = MediaColors.TextTertiary
-                )
-                
-                update.chapter.wordCount?.let { words ->
-                    Text("•", color = MediaColors.TextTertiary)
-                    Text(
-                        text = formatWordCount(words),
-                        style = MediaTypography.LabelSmall,
-                        color = MediaColors.TextTertiary
-                    )
-                }
-            }
-        }
-        
-        Icon(
-            imageVector = Icons.Default.ChevronRight,
-            contentDescription = "Media image",
-            tint = MediaColors.TextTertiary,
-            modifier = Modifier.size(20.dp)
-        )
-    }
-}
 
 // =============================================================================
 // BROWSE SOURCES PAGE
@@ -671,84 +285,6 @@ private fun BrowseSourcesPage(
     }
 }
 
-@Composable
-private fun SourceCard(
-    source: WebFictionSource,
-    onClick: () -> Unit
-) {
-    Surface(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(MediaCorners.MD))
-            .clickable(onClick = onClick),
-        color = MediaColors.BackgroundElevated
-    ) {
-        Row(
-            modifier = Modifier.padding(MediaSpacing.MD),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // Source icon
-            Surface(
-                shape = RoundedCornerShape(MediaCorners.SM),
-                color = source.color.copy(alpha = 0.2f),
-                modifier = Modifier.size(56.dp)
-            ) {
-                Box(contentAlignment = Alignment.Center) {
-                    Icon(
-                        imageVector = source.icon,
-                        contentDescription = "Media image",
-                        tint = source.color,
-                        modifier = Modifier.size(28.dp)
-                    )
-                }
-            }
-            
-            Spacer(modifier = Modifier.width(MediaSpacing.MD))
-            
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = source.displayName,
-                    style = MediaTypography.BodyLarge,
-                    color = MediaColors.TextPrimary,
-                    fontWeight = FontWeight.Medium
-                )
-                Text(
-                    text = source.description,
-                    style = MediaTypography.LabelSmall,
-                    color = MediaColors.TextSecondary,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis
-                )
-                
-                Spacer(modifier = Modifier.height(MediaSpacing.XS))
-                
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
-                ) {
-                    source.categories.take(3).forEach { category ->
-                        Surface(
-                            shape = RoundedCornerShape(MediaCorners.XS),
-                            color = MediaColors.BackgroundSurface
-                        ) {
-                            Text(
-                                text = category,
-                                style = MediaTypography.LabelSmall,
-                                color = MediaColors.TextTertiary,
-                                modifier = Modifier.padding(horizontal = MediaSpacing.XS, vertical = 1.dp)
-                            )
-                        }
-                    }
-                }
-            }
-            
-            Icon(
-                imageVector = Icons.Default.ChevronRight,
-                contentDescription = "Media image",
-                tint = MediaColors.TextTertiary
-            )
-        }
-    }
-}
 
 // =============================================================================
 // EMPTY STATE
@@ -816,98 +352,6 @@ data class WebFictionStory(
     val url: String
 )
 
-data class WebFictionChapter(
-    val id: String,
-    val storyId: String,
-    val title: String,
-    val chapterNumber: Int,
-    val wordCount: Long? = null,
-    val publishedDate: String,
-    val url: String
-)
-
-data class WebFictionUpdate(
-    val storyId: String,
-    val storyTitle: String,
-    val storyCoverUrl: String?,
-    val source: WebFictionSource,
-    val chapter: WebFictionChapter,
-    val timestamp: String,
-    val isRead: Boolean = false
-)
-
-data class WebFictionSource(
-    val id: String,
-    val displayName: String,
-    val description: String,
-    val icon: androidx.compose.ui.graphics.vector.ImageVector,
-    val color: Color,
-    val categories: List<String>,
-    val baseUrl: String,
-    val supportsDownload: Boolean = true
-) {
-    companion object {
-        val ROYAL_ROAD = WebFictionSource(
-            id = "royalroad",
-            displayName = "Royal Road",
-            description = "Home of web serials, LitRPG, and progression fantasy",
-            icon = Icons.Default.AutoStories,
-            color = Color(0xFF2E7D32),  // Green
-            categories = listOf("LitRPG", "Fantasy", "Sci-Fi"),
-            baseUrl = "https://www.royalroad.com"
-        )
-        
-        val AO3 = WebFictionSource(
-            id = "ao3",
-            displayName = "Archive of Our Own",
-            description = "Fan-created works across all fandoms",
-            icon = Icons.Default.Favorite,
-            color = Color(0xFF990000),  // AO3 Red
-            categories = listOf("Fanfiction", "Original Works"),
-            baseUrl = "https://archiveofourown.org"
-        )
-        
-        val FFN = WebFictionSource(
-            id = "ffn",
-            displayName = "FanFiction.Net",
-            description = "The largest archive of fanfiction stories",
-            icon = Icons.Default.MenuBook,
-            color = Color(0xFF0066CC),  // FFN Blue
-            categories = listOf("Anime", "Books", "Games", "Movies"),
-            baseUrl = "https://www.fanfiction.net"
-        )
-        
-        val WATTPAD = WebFictionSource(
-            id = "wattpad",
-            displayName = "Wattpad",
-            description = "Social storytelling platform",
-            icon = Icons.Default.Edit,
-            color = Color(0xFFFF500A),  // Wattpad Orange
-            categories = listOf("Romance", "Teen Fiction", "Werewolf"),
-            baseUrl = "https://www.wattpad.com"
-        )
-        
-        val SCRIBBLEHUB = WebFictionSource(
-            id = "scribblehub",
-            displayName = "ScribbleHub",
-            description = "Reading and writing community for web novels",
-            icon = Icons.Default.Create,
-            color = Color(0xFF9C27B0),  // Purple
-            categories = listOf("Fantasy", "Isekai", "LitRPG"),
-            baseUrl = "https://www.scribblehub.com"
-        )
-        
-        val SPACEBATTLES = WebFictionSource(
-            id = "spacebattles",
-            displayName = "SpaceBattles",
-            description = "Forums for creative writing and discussion",
-            icon = Icons.Default.RocketLaunch,
-            color = Color(0xFF455A64),  // Blue Grey
-            categories = listOf("Crossover", "Alt-History", "Sci-Fi"),
-            baseUrl = "https://forums.spacebattles.com"
-        )
-    }
-}
 
 enum class StoryStatus(val displayName: String) {
     ONGOING("Ongoing"),
