+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/webfiction/DownloadTab.kt
@@ -1,0 +1,387 @@
+package com.universalmedialibrary.ui.webfiction
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.automirrored.filled.ArrowForward
+import androidx.compose.material.icons.automirrored.filled.LibraryBooks
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.NavController
+import com.universalmedialibrary.services.webfiction.*
+import com.universalmedialibrary.ui.components.PinAccessDialog
+import com.universalmedialibrary.ui.theme.CleverFerretTheme
+import com.universalmedialibrary.ui.theme.ThemePalette
+import java.util.Locale
+import com.universalmedialibrary.ui.fanfiction.FanfictionViewModel
+import com.universalmedialibrary.data.local.entity.FanfictionStoryEntity
+import com.universalmedialibrary.ui.fanfiction.FilterType
+import com.universalmedialibrary.ui.icons.PhosphorIcons
+
+/**
+ * Unified Fanfiction Hub - All-in-one interface for fanfiction discovery, download, and management.
+ * 
+ * Streamlined experience combining:
+ * - Site selection & Tag-based browsing
+ * - Direct story download
+ * - Library management (My Library)
+ * - Update checker
+ * 
+ * No more jumping between screens! Reading is handled by the separate eReader.
+ */
+
+// Tab indices for the unified fanfiction hub
+/**
+ * Download Tab - Direct URL download interface with supported sites info
+ */
+@Composable
+internal fun DownloadTab(
+    downloadState: FanfictionDownloaderUiState,
+    downloadViewModel: FanfictionDownloaderViewModel,
+    adultSitesEnabled: Boolean,
+    navController: NavController
+) {
+    var url by remember { mutableStateOf("") }
+    
+    LazyColumn(
+        modifier = Modifier.fillMaxSize(),
+        contentPadding = PaddingValues(16.dp),
+        verticalArrangement = Arrangement.spacedBy(16.dp)
+    ) {
+        // URL Input Card
+        item {
+            Card(
+                modifier = Modifier.fillMaxWidth(),
+                colors = CardDefaults.cardColors(
+                    containerColor = MaterialTheme.colorScheme.primaryContainer
+                )
+            ) {
+                Column(
+                    modifier = Modifier.padding(16.dp),
+                    verticalArrangement = Arrangement.spacedBy(12.dp)
+                ) {
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically
+                    ) {
+                        Icon(
+                            Icons.Default.Download,
+                            contentDescription = "Media image",
+                            tint = MaterialTheme.colorScheme.primary,
+                            modifier = Modifier.size(32.dp)
+                        )
+                        Spacer(modifier = Modifier.width(12.dp))
+                        Column {
+                            Text(
+                                text = "Download Story",
+                                style = MaterialTheme.typography.titleMedium,
+                                fontWeight = FontWeight.Bold
+                            )
+                            Text(
+                                text = "Paste a story URL to download as EPUB",
+                                style = MaterialTheme.typography.bodySmall,
+                                color = MaterialTheme.colorScheme.onPrimaryContainer
+                            )
+                        }
+                    }
+                    
+                    OutlinedTextField(
+                        value = url,
+                        onValueChange = { url = it },
+                        modifier = Modifier.fillMaxWidth(),
+                        label = { Text("Story URL") },
+                        placeholder = { Text("https://archiveofourown.org/works/...") },
+                        leadingIcon = {
+                            Icon(Icons.Default.Link, contentDescription = "Link")
+                        },
+                        singleLine = true,
+                        enabled = !downloadState.isLoading
+                    )
+                    
+                    Button(
+                        onClick = {
+                            if (url.isNotBlank()) {
+                                downloadViewModel.downloadFromUrl(url)
+                            }
+                        },
+                        modifier = Modifier.fillMaxWidth(),
+                        enabled = url.isNotBlank() && !downloadState.isLoading
+                    ) {
+                        if (downloadState.isLoading) {
+                            CircularProgressIndicator(
+                                modifier = Modifier.size(20.dp),
+                                color = MaterialTheme.colorScheme.onPrimary,
+                                strokeWidth = 2.dp
+                            )
+                            Spacer(modifier = Modifier.width(8.dp))
+                            Text("Downloading...")
+                        } else {
+                            Icon(Icons.Default.Download, contentDescription = "Download")
+                            Spacer(modifier = Modifier.width(8.dp))
+                            Text("Download as EPUB")
+                        }
+                    }
+                }
+            }
+        }
+        
+        // Progress Info
+        if (downloadState.isLoading) {
+            item {
+                Card(
+                    modifier = Modifier.fillMaxWidth(),
+                    colors = CardDefaults.cardColors(
+                        containerColor = MaterialTheme.colorScheme.surfaceVariant
+                    )
+                ) {
+                    Column(
+                        modifier = Modifier.padding(16.dp),
+                        verticalArrangement = Arrangement.spacedBy(8.dp),
+                        horizontalAlignment = Alignment.CenterHorizontally
+                    ) {
+                        LinearProgressIndicator(modifier = Modifier.fillMaxWidth())
+                        Text(
+                            text = downloadState.progressMessage ?: "Fetching story...",
+                            style = MaterialTheme.typography.bodyMedium,
+                            color = MaterialTheme.colorScheme.onSurfaceVariant
+                        )
+                    }
+                }
+            }
+        }
+        
+        // Success Result
+        downloadState.result?.let { result ->
+            item {
+                Card(
+                    modifier = Modifier.fillMaxWidth(),
+                    colors = CardDefaults.cardColors(
+                        containerColor = MaterialTheme.colorScheme.tertiaryContainer
+                    )
+                ) {
+                    Column(
+                        modifier = Modifier.padding(16.dp),
+                        verticalArrangement = Arrangement.spacedBy(12.dp)
+                    ) {
+                        Row(
+                            verticalAlignment = Alignment.CenterVertically
+                        ) {
+                            Icon(
+                                Icons.Default.CheckCircle,
+                                contentDescription = "Media image",
+                                tint = MaterialTheme.colorScheme.tertiary,
+                                modifier = Modifier.size(32.dp)
+                            )
+                            Spacer(modifier = Modifier.width(12.dp))
+                            Text(
+                                text = "Download Complete!",
+                                style = MaterialTheme.typography.titleLarge,
+                                fontWeight = FontWeight.Bold,
+                                color = MaterialTheme.colorScheme.onTertiaryContainer
+                            )
+                        }
+                        
+                        HorizontalDivider()
+                        
+                        Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
+                            DownloadInfoRow("Title", result.title)
+                            DownloadInfoRow("Author", result.author)
+                            DownloadInfoRow("Chapters", result.chapters.toString())
+                        }
+                        
+                        Row(
+                            modifier = Modifier.fillMaxWidth(),
+                            horizontalArrangement = Arrangement.spacedBy(8.dp)
+                        ) {
+                            OutlinedButton(
+                                onClick = { downloadViewModel.openEpub() },
+                                modifier = Modifier.weight(1f)
+                            ) {
+                                Icon(Icons.Default.Book, contentDescription = "Book")
+                                Spacer(modifier = Modifier.width(4.dp))
+                                Text("Read")
+                            }
+                            Button(
+                                onClick = { downloadViewModel.shareEpub() },
+                                modifier = Modifier.weight(1f)
+                            ) {
+                                Icon(Icons.Default.Share, contentDescription = "Share")
+                                Spacer(modifier = Modifier.width(4.dp))
+                                Text("Share")
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        
+        // Error Display
+        downloadState.error?.let { error ->
+            item {
+                Card(
+                    modifier = Modifier.fillMaxWidth(),
+                    colors = CardDefaults.cardColors(
+                        containerColor = MaterialTheme.colorScheme.errorContainer
+                    )
+                ) {
+                    Row(
+                        modifier = Modifier.padding(16.dp),
+                        verticalAlignment = Alignment.CenterVertically
+                    ) {
+                        Icon(
+                            Icons.Default.Error,
+                            contentDescription = "Media image",
+                            tint = MaterialTheme.colorScheme.error
+                        )
+                        Spacer(modifier = Modifier.width(12.dp))
+                        Column(modifier = Modifier.weight(1f)) {
+                            Text(
+                                text = "Download Failed",
+                                style = MaterialTheme.typography.titleMedium,
+                                fontWeight = FontWeight.Medium,
+                                color = MaterialTheme.colorScheme.onErrorContainer
+                            )
+                            Text(
+                                text = error,
+                                style = MaterialTheme.typography.bodyMedium,
+                                color = MaterialTheme.colorScheme.onErrorContainer
+                            )
+                        }
+                    }
+                }
+            }
+        }
+        
+        // Supported Sites Card
+        item {
+            Card(modifier = Modifier.fillMaxWidth()) {
+                Column(
+                    modifier = Modifier.padding(16.dp),
+                    verticalArrangement = Arrangement.spacedBy(12.dp)
+                ) {
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically
+                    ) {
+                        Icon(
+                            Icons.Default.Language,
+                            contentDescription = "Media image",
+                            tint = MaterialTheme.colorScheme.primary
+                        )
+                        Spacer(modifier = Modifier.width(8.dp))
+                        Text(
+                            text = "Supported Sites",
+                            style = MaterialTheme.typography.titleMedium,
+                            fontWeight = FontWeight.Medium
+                        )
+                    }
+                    
+                    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
+                        SupportedSiteRow("Archive of Our Own", "archiveofourown.org", "✓ Full support")
+                        SupportedSiteRow("FanFiction.Net", "fanfiction.net", "✓ Multi-chapter")
+                        SupportedSiteRow("Royal Road", "royalroad.com", "✓ Web fiction")
+                        SupportedSiteRow("Wattpad", "wattpad.com", "✓ Basic support")
+                        if (adultSitesEnabled) {
+                            SupportedSiteRow("Metabods", "metabods.com", "✓ Adult content")
+                            SupportedSiteRow("Literotica", "literotica.com", "✓ Adult content")
+                        }
+                        SupportedSiteRow("Other Sites", "Generic", "Limited support")
+                    }
+                }
+            }
+        }
+        
+        // Tips Card
+        item {
+            Card(
+                modifier = Modifier.fillMaxWidth(),
+                colors = CardDefaults.cardColors(
+                    containerColor = MaterialTheme.colorScheme.surfaceVariant
+                )
+            ) {
+                Column(
+                    modifier = Modifier.padding(16.dp),
+                    verticalArrangement = Arrangement.spacedBy(8.dp)
+                ) {
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically
+                    ) {
+                        Icon(
+                            Icons.Default.Lightbulb,
+                            contentDescription = "Media image",
+                            tint = MaterialTheme.colorScheme.tertiary
+                        )
+                        Spacer(modifier = Modifier.width(8.dp))
+                        Text(
+                            text = "Tips",
+                            style = MaterialTheme.typography.titleSmall,
+                            fontWeight = FontWeight.Medium
+                        )
+                    }
+                    Text(
+                        text = "• Use the Discover tab to browse stories by tags\n" +
+                               "• Downloaded stories appear in Library tab\n" +
+                               "• EPUBs are saved to your device for offline reading\n" +
+                               "• Large stories with many chapters may take longer",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+        }
+        
+        // Web Comics Card
+        item {
+            Card(
+                modifier = Modifier.fillMaxWidth(),
+                colors = CardDefaults.cardColors(
+                    containerColor = MaterialTheme.colorScheme.secondaryContainer
+                ),
+                onClick = { navController.navigate("web_comic_downloader") }
+            ) {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(16.dp),
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Icon(
+                        Icons.Default.PhotoLibrary,
+                        contentDescription = "Media image",
+                        tint = MaterialTheme.colorScheme.secondary,
+                        modifier = Modifier.size(32.dp)
+                    )
+                    Spacer(modifier = Modifier.width(12.dp))
+                    Column(modifier = Modifier.weight(1f)) {
+                        Text(
+                            text = "Web Comic Downloader",
+                            style = MaterialTheme.typography.titleMedium,
+                            fontWeight = FontWeight.Bold
+                        )
+                        Text(
+                            text = "Search ComicVine and download web comics",
+                            style = MaterialTheme.typography.bodySmall,
+                            color = MaterialTheme.colorScheme.onSecondaryContainer
+                        )
+                    }
+                    Icon(
+                        Icons.AutoMirrored.Filled.ArrowForward,
+                        contentDescription = "Open",
+                        tint = MaterialTheme.colorScheme.secondary
+                    )
+                }
+            }
+        }
+    }
+}
