+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/webfiction/UnifiedFanfictionHubScreen.kt
@@ -222,633 +222,13 @@ fun UnifiedFanfictionHubScreen(
         }
 }
 
-@Composable
-fun FanfictionLibraryTab(
-    viewModel: FanfictionViewModel,
-    onStoryClick: (FanfictionStoryEntity) -> Unit
-) {
-    val stories by viewModel.allStories.collectAsState()
-    val selectedFilter by viewModel.selectedFilter.collectAsState()
-    val updateStatus by viewModel.updateStatus.collectAsState()
-    
-    var showFilterMenu by remember { mutableStateOf(false) }
-    var searchQuery by remember { mutableStateOf("") }
-    
-    val filteredStories = remember(stories, selectedFilter, searchQuery) {
-        stories.filter { story ->
-            val matchesSearch = searchQuery.isBlank() ||
-                story.title.contains(searchQuery, ignoreCase = true) ||
-                story.author.contains(searchQuery, ignoreCase = true) ||
-                story.summary.contains(searchQuery, ignoreCase = true)
-            
-            val matchesFilter = when (selectedFilter) {
-                FilterType.ALL -> true
-                FilterType.IN_PROGRESS -> story.status == "IN_PROGRESS"
-                FilterType.COMPLETE -> story.status == "COMPLETE"
-                FilterType.AO3 -> story.sourceSite == "Archive of Our Own"
-                FilterType.FFN -> story.sourceSite == "FanFiction.Net"
-                FilterType.ROYAL_ROAD -> story.sourceSite == "Royal Road"
-            }
-            
-            matchesSearch && matchesFilter
-        }
-    }
-
-    Column(modifier = Modifier.fillMaxSize()) {
-        // Filter and Search Row
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            OutlinedTextField(
-                value = searchQuery,
-                onValueChange = { searchQuery = it },
-                placeholder = { Text("Search library...") },
-                leadingIcon = { Icon(Icons.Default.Search, "Search") },
-                modifier = Modifier
-                    .weight(1f)
-                    .padding(end = 8.dp),
-                singleLine = true
-            )
-            
-            IconButton(onClick = { showFilterMenu = true }) {
-                Icon(Icons.Default.FilterList, "Filter")
-            }
-            
-            DropdownMenu(
-                expanded = showFilterMenu,
-                onDismissRequest = { showFilterMenu = false }
-            ) {
-                FilterType.values().forEach { filter ->
-                    DropdownMenuItem(
-                        text = { Text(filter.name.replace("_", " ")) },
-                        onClick = {
-                            viewModel.setFilter(filter)
-                            showFilterMenu = false
-                        },
-                        leadingIcon = {
-                            if (filter == selectedFilter) {
-                                Icon(Icons.Default.Check, null)
-                            }
-                        }
-                    )
-                }
-            }
-        }
-        
-        // Update status banner
-        updateStatus?.let { status ->
-            Surface(
-                color = MaterialTheme.colorScheme.primaryContainer,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Row(
-                    modifier = Modifier.padding(12.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        status,
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onPrimaryContainer,
-                        modifier = Modifier.weight(1f)
-                    )
-                    IconButton(onClick = { viewModel.clearUpdateStatus() }) {
-                        Icon(Icons.Default.Close, "Dismiss")
-                    }
-                }
-            }
-        }
-
-        if (filteredStories.isEmpty()) {
-            Box(
-                modifier = Modifier.fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                Text(
-                    if (stories.isNotEmpty()) "No stories match your filter" else "Your library is empty. Go to 'Discover' to find stories!",
-                    style = MaterialTheme.typography.bodyLarge,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    textAlign = TextAlign.Center,
-                    modifier = Modifier.padding(32.dp)
-                )
-            }
-        } else {
-            LazyColumn(
-                contentPadding = PaddingValues(16.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                items(
-                    items = filteredStories,
-                    key = { it.id }
-                ) { story ->
-                    LibraryStoryCard(
-                        story = story,
-                        onClick = { onStoryClick(story) },
-                        onUpdateClick = { viewModel.updateStory(story.id) },
-                        onDeleteClick = { viewModel.deleteStory(story) }
-                    )
-                }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-internal fun LibraryStoryCard(
-    story: FanfictionStoryEntity,
-    onClick: () -> Unit,
-    onUpdateClick: () -> Unit,
-    onDeleteClick: () -> Unit
-) {
-    var showMenu by rememberSaveable(story.id) { mutableStateOf(false) }
-    
-    Card(
-        onClick = onClick,
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp)
-        ) {
-            // Title and menu
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    story.title,
-                    style = MaterialTheme.typography.titleMedium,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.weight(1f)
-                )
-                
-                Box {
-                    IconButton(onClick = { showMenu = true }) {
-                        Icon(Icons.Default.MoreVert, "Menu")
-                    }
-                    
-                    DropdownMenu(
-                        expanded = showMenu,
-                        onDismissRequest = { showMenu = false }
-                    ) {
-                        if (story.status == "IN_PROGRESS") {
-                            DropdownMenuItem(
-                                text = { Text("Check for Updates") },
-                                onClick = {
-                                    onUpdateClick()
-                                    showMenu = false
-                                },
-                                leadingIcon = {
-                                    Icon(Icons.Default.Refresh, null)
-                                }
-                            )
-                        }
-                        DropdownMenuItem(
-                            text = { Text("Delete") },
-                            onClick = {
-                                onDeleteClick()
-                                showMenu = false
-                            },
-                            leadingIcon = {
-                                Icon(Icons.Default.Delete, null)
-                            }
-                        )
-                    }
-                }
-            }
-            
-            // Author
-            Text(
-                "by ${story.author}",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            
-            Spacer(Modifier.height(8.dp))
-            
-            // Summary
-            Text(
-                story.summary,
-                style = MaterialTheme.typography.bodySmall,
-                maxLines = 3,
-                overflow = TextOverflow.Ellipsis,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            
-            Spacer(Modifier.height(12.dp))
-            
-            // Stats
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                Text(
-                    "${story.chapterCount} ch",
-                    style = MaterialTheme.typography.bodySmall
-                )
-                Text(
-                    formatNumber(story.wordCount),
-                    style = MaterialTheme.typography.bodySmall
-                )
-                Text(
-                    story.sourceSite,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.primary
-                )
-            }
-        }
-    }
-}
 
 
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun UnifiedContent(
-    uiState: UniversalTagBrowserUiState,
-    downloadState: FanfictionDownloaderUiState,
-    viewModel: UniversalTagBrowserViewModel,
-    downloadViewModel: FanfictionDownloaderViewModel,
-    adultSitesEnabled: Boolean
-) {
-    LazyColumn(
-        modifier = Modifier.fillMaxSize(),
-        contentPadding = PaddingValues(12.dp),
-        verticalArrangement = Arrangement.spacedBy(12.dp)
-    ) {
-        // Site Selector Card
-        item {
-            SiteSelectorCard(
-                selectedSite = uiState.selectedSite,
-                adultSitesEnabled = adultSitesEnabled,
-                onSiteSelected = { viewModel.selectSite(it) }
-            )
-        }
-        
-        // Only show content if a site is selected
-        if (uiState.selectedSite != null) {
-            // Selected tags display
-            if (uiState.selectedTags.isNotEmpty()) {
-                item {
-                    SelectedTagsCard(
-                        selectedTags = uiState.selectedTags,
-                        onTagRemove = { viewModel.toggleTag(it) },
-                        onClearAll = { viewModel.clearTags() }
-                    )
-                }
-            }
-            
-            // Available tags
-            item {
-                AvailableTagsCard(
-                    tags = uiState.tags,
-                    selectedTags = uiState.selectedTags,
-                    onTagToggle = { viewModel.toggleTag(it.name) }
-                )
-            }
-            
-            // Search button
-            if (uiState.selectedTags.isNotEmpty()) {
-                item {
-                    Button(
-                        onClick = { viewModel.browseStories() },
-                        modifier = Modifier.fillMaxWidth(),
-                        enabled = !uiState.isLoadingStories
-                    ) {
-                        Icon(Icons.Default.Search, null)
-                        Spacer(modifier = Modifier.width(8.dp))
-                        Text("Search Stories")
-                    }
-                }
-            }
-            
-            // Search results
-            if (uiState.isLoadingStories) {
-                item {
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(32.dp),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        CircularProgressIndicator()
-                    }
-                }
-            }
-            
-            uiState.searchResult?.let { result ->
-                if (result.stories.isNotEmpty()) {
-                    item {
-                        Text(
-                            "${result.stories.size} Stories Found",
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-                    
-                    items(
-                        items = result.stories,
-                        key = { story -> story.id }
-                    ) { story ->
-                        StoryCard(
-                            story = story,
-                            onDownload = { 
-                                story.url?.let { downloadViewModel.downloadFromUrl(it) }
-                            },
-                            isDownloading = downloadState.isLoading && downloadState.url == story.url
-                        )
-                    }
-                } else if (!uiState.isLoadingStories) {
-                    item {
-                        EmptyResultsCard()
-                    }
-                }
-            }
-        }
-    }
-}
 
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun SiteSelectorCard(
-    selectedSite: WebFictionSiteType?,
-    adultSitesEnabled: Boolean,
-    onSiteSelected: (WebFictionSiteType) -> Unit
-) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.secondaryContainer
-        )
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = "Select a Site",
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Bold
-                )
-                if (selectedSite != null) {
-                    FilterChip(
-                        selected = false,
-                        onClick = { onSiteSelected(selectedSite) },
-                        label = { Text(selectedSite.displayName) },
-                        leadingIcon = {
-                            Icon(Icons.Default.Check, null, modifier = Modifier.size(18.dp))
-                        }
-                    )
-                }
-            }
-            
-            LazyRow(
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                items(
-                    items = WebFictionSiteType.entries.toTypedArray(),
-                    key = { siteType -> siteType.name }
-                ) { siteType ->
-                    val isAdult = siteType.isAdultSite()
-                    val isEnabled = !isAdult || adultSitesEnabled
-                    
-                    FilterChip(
-                        selected = selectedSite == siteType,
-                        onClick = { if (isEnabled) onSiteSelected(siteType) },
-                        label = { Text(siteType.displayName) },
-                        enabled = isEnabled,
-                        leadingIcon = if (selectedSite == siteType) {
-                            { Icon(Icons.Default.Check, null, modifier = Modifier.size(18.dp)) }
-                        } else null
-                    )
-                }
-            }
-            
-            if (!adultSitesEnabled) {
-                Text(
-                    "Some sites require enabling Adult Content in Parental Controls",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-    }
-}
 
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun SelectedTagsCard(
-    selectedTags: List<String>,
-    onTagRemove: (String) -> Unit,
-    onClearAll: () -> Unit
-) {
-    Card(
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.tertiaryContainer
-        )
-    ) {
-        Column(modifier = Modifier.padding(12.dp)) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    "Selected Tags (${selectedTags.size})",
-                    style = MaterialTheme.typography.titleSmall,
-                    fontWeight = FontWeight.Bold
-                )
-                TextButton(onClick = onClearAll) {
-                    Text("Clear All")
-                }
-            }
-            FlowRow(
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                selectedTags.forEach { tag ->
-                    FilterChip(
-                        selected = true,
-                        onClick = { onTagRemove(tag) },
-                        label = { Text(tag) },
-                        trailingIcon = {
-                            Icon(
-                                Icons.Default.Close,
-                                null,
-                                modifier = Modifier.size(16.dp)
-                            )
-                        }
-                    )
-                }
-            }
-        }
-    }
-}
 
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun AvailableTagsCard(
-    tags: List<WebFictionTag>,
-    selectedTags: List<String>,
-    onTagToggle: (WebFictionTag) -> Unit
-) {
-    Card {
-        Column(modifier = Modifier.padding(12.dp)) {
-            Text(
-                "Browse Tags",
-                style = MaterialTheme.typography.titleSmall,
-                fontWeight = FontWeight.Bold
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            
-            if (tags.isEmpty()) {
-                Text(
-                    "Select a site to see available tags",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            } else {
-                FlowRow(
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    tags.take(50).forEach { tag ->
-                        FilterChip(
-                            selected = tag.name in selectedTags,
-                            onClick = { onTagToggle(tag) },
-                            label = { Text(tag.name) }
-                        )
-                    }
-                }
-                
-                if (tags.size > 50) {
-                    Text(
-                        "... and ${tags.size - 50} more tags",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        modifier = Modifier.padding(top = 8.dp)
-                    )
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun StoryCard(
-    story: WebFictionStory,
-    onDownload: () -> Unit,
-    isDownloading: Boolean
-) {
-    var showDetails by remember { mutableStateOf(false) }
-    
-    ElevatedCard(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { showDetails = !showDetails }
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            // Title and author
-            Text(
-                text = story.title,
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Bold
-            )
-            Text(
-                text = "by ${story.author}",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            
-            // Stats row
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 8.dp),
-                horizontalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                story.wordCount?.let {
-                    Text(
-                        "📝 ${formatNumber(it.toInt())} words",
-                        style = MaterialTheme.typography.bodySmall
-                    )
-                }
-                story.chapterCount?.let {
-                    Text(
-                        "📚 $it chapters",
-                        style = MaterialTheme.typography.bodySmall
-                    )
-                }
-                Text(
-                    if (story.status == StoryStatus.COMPLETED) "✅ Complete" else "📝 In Progress",
-                    style = MaterialTheme.typography.bodySmall
-                )
-            }
-            
-            // Expandable description
-            AnimatedVisibility(visible = showDetails) {
-                Column {
-                    HorizontalDivider(modifier = Modifier.padding(vertical = 8.dp))
-                    story.description?.let {
-                        Text(
-                            text = it,
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                    
-                    // Tags
-                    if (story.tags.isNotEmpty()) {
-                        Spacer(modifier = Modifier.height(8.dp))
-                        Text(
-                            "Tags: ${story.tags.joinToString(", ")}",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                }
-            }
-            
-            // Action buttons
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Button(
-                    onClick = onDownload,
-                    modifier = Modifier.weight(1f),
-                    enabled = !isDownloading
-                ) {
-                    Icon(Icons.Default.Download, null, modifier = Modifier.size(18.dp))
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Text(if (isDownloading) "Downloading..." else "Download")
-                }
-                OutlinedButton(
-                    onClick = { showDetails = !showDetails },
-                    modifier = Modifier.weight(1f)
-                ) {
-                    Icon(
-                        if (showDetails) Icons.Default.ExpandLess else Icons.Default.ExpandMore,
-                        null,
-                        modifier = Modifier.size(18.dp)
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Text(if (showDetails) "Less" else "More")
-                }
-            }
-        }
-    }
-}
+
 
 @Composable
 private fun EmptyResultsCard() {
@@ -929,347 +309,6 @@ private fun QuickDownloadDialog(
     )
 }
 
-/**
- * Download Tab - Direct URL download interface with supported sites info
- */
-@Composable
-private fun DownloadTab(
-    downloadState: FanfictionDownloaderUiState,
-    downloadViewModel: FanfictionDownloaderViewModel,
-    adultSitesEnabled: Boolean,
-    navController: NavController
-) {
-    var url by remember { mutableStateOf("") }
-    
-    LazyColumn(
-        modifier = Modifier.fillMaxSize(),
-        contentPadding = PaddingValues(16.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        // URL Input Card
-        item {
-            Card(
-                modifier = Modifier.fillMaxWidth(),
-                colors = CardDefaults.cardColors(
-                    containerColor = MaterialTheme.colorScheme.primaryContainer
-                )
-            ) {
-                Column(
-                    modifier = Modifier.padding(16.dp),
-                    verticalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Icon(
-                            Icons.Default.Download,
-                            contentDescription = "Media image",
-                            tint = MaterialTheme.colorScheme.primary,
-                            modifier = Modifier.size(32.dp)
-                        )
-                        Spacer(modifier = Modifier.width(12.dp))
-                        Column {
-                            Text(
-                                text = "Download Story",
-                                style = MaterialTheme.typography.titleMedium,
-                                fontWeight = FontWeight.Bold
-                            )
-                            Text(
-                                text = "Paste a story URL to download as EPUB",
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.onPrimaryContainer
-                            )
-                        }
-                    }
-                    
-                    OutlinedTextField(
-                        value = url,
-                        onValueChange = { url = it },
-                        modifier = Modifier.fillMaxWidth(),
-                        label = { Text("Story URL") },
-                        placeholder = { Text("https://archiveofourown.org/works/...") },
-                        leadingIcon = {
-                            Icon(Icons.Default.Link, contentDescription = "Link")
-                        },
-                        singleLine = true,
-                        enabled = !downloadState.isLoading
-                    )
-                    
-                    Button(
-                        onClick = {
-                            if (url.isNotBlank()) {
-                                downloadViewModel.downloadFromUrl(url)
-                            }
-                        },
-                        modifier = Modifier.fillMaxWidth(),
-                        enabled = url.isNotBlank() && !downloadState.isLoading
-                    ) {
-                        if (downloadState.isLoading) {
-                            CircularProgressIndicator(
-                                modifier = Modifier.size(20.dp),
-                                color = MaterialTheme.colorScheme.onPrimary,
-                                strokeWidth = 2.dp
-                            )
-                            Spacer(modifier = Modifier.width(8.dp))
-                            Text("Downloading...")
-                        } else {
-                            Icon(Icons.Default.Download, contentDescription = "Download")
-                            Spacer(modifier = Modifier.width(8.dp))
-                            Text("Download as EPUB")
-                        }
-                    }
-                }
-            }
-        }
-        
-        // Progress Info
-        if (downloadState.isLoading) {
-            item {
-                Card(
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.surfaceVariant
-                    )
-                ) {
-                    Column(
-                        modifier = Modifier.padding(16.dp),
-                        verticalArrangement = Arrangement.spacedBy(8.dp),
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-                        LinearProgressIndicator(modifier = Modifier.fillMaxWidth())
-                        Text(
-                            text = downloadState.progressMessage ?: "Fetching story...",
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                }
-            }
-        }
-        
-        // Success Result
-        downloadState.result?.let { result ->
-            item {
-                Card(
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.tertiaryContainer
-                    )
-                ) {
-                    Column(
-                        modifier = Modifier.padding(16.dp),
-                        verticalArrangement = Arrangement.spacedBy(12.dp)
-                    ) {
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Icon(
-                                Icons.Default.CheckCircle,
-                                contentDescription = "Media image",
-                                tint = MaterialTheme.colorScheme.tertiary,
-                                modifier = Modifier.size(32.dp)
-                            )
-                            Spacer(modifier = Modifier.width(12.dp))
-                            Text(
-                                text = "Download Complete!",
-                                style = MaterialTheme.typography.titleLarge,
-                                fontWeight = FontWeight.Bold,
-                                color = MaterialTheme.colorScheme.onTertiaryContainer
-                            )
-                        }
-                        
-                        HorizontalDivider()
-                        
-                        Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
-                            DownloadInfoRow("Title", result.title)
-                            DownloadInfoRow("Author", result.author)
-                            DownloadInfoRow("Chapters", result.chapters.toString())
-                        }
-                        
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.spacedBy(8.dp)
-                        ) {
-                            OutlinedButton(
-                                onClick = { downloadViewModel.openEpub() },
-                                modifier = Modifier.weight(1f)
-                            ) {
-                                Icon(Icons.Default.Book, contentDescription = "Book")
-                                Spacer(modifier = Modifier.width(4.dp))
-                                Text("Read")
-                            }
-                            Button(
-                                onClick = { downloadViewModel.shareEpub() },
-                                modifier = Modifier.weight(1f)
-                            ) {
-                                Icon(Icons.Default.Share, contentDescription = "Share")
-                                Spacer(modifier = Modifier.width(4.dp))
-                                Text("Share")
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        
-        // Error Display
-        downloadState.error?.let { error ->
-            item {
-                Card(
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.errorContainer
-                    )
-                ) {
-                    Row(
-                        modifier = Modifier.padding(16.dp),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Icon(
-                            Icons.Default.Error,
-                            contentDescription = "Media image",
-                            tint = MaterialTheme.colorScheme.error
-                        )
-                        Spacer(modifier = Modifier.width(12.dp))
-                        Column(modifier = Modifier.weight(1f)) {
-                            Text(
-                                text = "Download Failed",
-                                style = MaterialTheme.typography.titleMedium,
-                                fontWeight = FontWeight.Medium,
-                                color = MaterialTheme.colorScheme.onErrorContainer
-                            )
-                            Text(
-                                text = error,
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.onErrorContainer
-                            )
-                        }
-                    }
-                }
-            }
-        }
-        
-        // Supported Sites Card
-        item {
-            Card(modifier = Modifier.fillMaxWidth()) {
-                Column(
-                    modifier = Modifier.padding(16.dp),
-                    verticalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Icon(
-                            Icons.Default.Language,
-                            contentDescription = "Media image",
-                            tint = MaterialTheme.colorScheme.primary
-                        )
-                        Spacer(modifier = Modifier.width(8.dp))
-                        Text(
-                            text = "Supported Sites",
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.Medium
-                        )
-                    }
-                    
-                    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                        SupportedSiteRow("Archive of Our Own", "archiveofourown.org", "✓ Full support")
-                        SupportedSiteRow("FanFiction.Net", "fanfiction.net", "✓ Multi-chapter")
-                        SupportedSiteRow("Royal Road", "royalroad.com", "✓ Web fiction")
-                        SupportedSiteRow("Wattpad", "wattpad.com", "✓ Basic support")
-                        if (adultSitesEnabled) {
-                            SupportedSiteRow("Metabods", "metabods.com", "✓ Adult content")
-                            SupportedSiteRow("Literotica", "literotica.com", "✓ Adult content")
-                        }
-                        SupportedSiteRow("Other Sites", "Generic", "Limited support")
-                    }
-                }
-            }
-        }
-        
-        // Tips Card
-        item {
-            Card(
-                modifier = Modifier.fillMaxWidth(),
-                colors = CardDefaults.cardColors(
-                    containerColor = MaterialTheme.colorScheme.surfaceVariant
-                )
-            ) {
-                Column(
-                    modifier = Modifier.padding(16.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Icon(
-                            Icons.Default.Lightbulb,
-                            contentDescription = "Media image",
-                            tint = MaterialTheme.colorScheme.tertiary
-                        )
-                        Spacer(modifier = Modifier.width(8.dp))
-                        Text(
-                            text = "Tips",
-                            style = MaterialTheme.typography.titleSmall,
-                            fontWeight = FontWeight.Medium
-                        )
-                    }
-                    Text(
-                        text = "• Use the Discover tab to browse stories by tags\n" +
-                               "• Downloaded stories appear in Library tab\n" +
-                               "• EPUBs are saved to your device for offline reading\n" +
-                               "• Large stories with many chapters may take longer",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-        }
-        
-        // Web Comics Card
-        item {
-            Card(
-                modifier = Modifier.fillMaxWidth(),
-                colors = CardDefaults.cardColors(
-                    containerColor = MaterialTheme.colorScheme.secondaryContainer
-                ),
-                onClick = { navController.navigate("web_comic_downloader") }
-            ) {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(16.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(
-                        Icons.Default.PhotoLibrary,
-                        contentDescription = "Media image",
-                        tint = MaterialTheme.colorScheme.secondary,
-                        modifier = Modifier.size(32.dp)
-                    )
-                    Spacer(modifier = Modifier.width(12.dp))
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(
-                            text = "Web Comic Downloader",
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.Bold
-                        )
-                        Text(
-                            text = "Search ComicVine and download web comics",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSecondaryContainer
-                        )
-                    }
-                    Icon(
-                        Icons.AutoMirrored.Filled.ArrowForward,
-                        contentDescription = "Open",
-                        tint = MaterialTheme.colorScheme.secondary
-                    )
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun DownloadInfoRow(label: String, value: String) {
