+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/webfiction/WebFictionManagerScreen.kt
@@ -410,408 +410,9 @@ fun WebFictionManagerScreen(
     }
 }
 
-@Composable
-fun WebFictionStoryCard(
-    story: WebFictionStory,
-    hasUpdates: Boolean,
-    onStoryClick: () -> Unit,
-    onUpdateClick: () -> Unit,
-    onDownloadClick: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onStoryClick),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surface
-        )
-    ) {
-        Row(
-            modifier = Modifier.padding(16.dp)
-        ) {
-            // Cover image
-            AsyncImage(
-                    
-                model = story.coverUrl ?: "https://via.placeholder.com/80x120/2d3136/e5a00d?text=📖",
-                contentDescription = "Story Cover",
-                modifier = Modifier
-                    .size(60.dp, 90.dp)
-                    .clip(MaterialTheme.shapes.small)
-                    .background(MaterialTheme.colorScheme.surfaceVariant),
-                contentScale = ContentScale.Crop
-            )
 
-            Spacer(modifier = Modifier.width(16.dp))
 
-            // Story info
-            Column(
-                modifier = Modifier.weight(1f)
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = story.title,
-                        style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.Medium,
-                        maxLines = 2,
-                        overflow = TextOverflow.Ellipsis,
-                        modifier = Modifier.weight(1f)
-                    )
 
-                    if (hasUpdates) {
-                        Icon(
-                            Icons.Default.NewReleases,
-                            contentDescription = "Has Updates",
-                            tint = MaterialTheme.colorScheme.primary,
-                            modifier = Modifier.size(20.dp)
-                        )
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(4.dp))
-
-                Text(
-                    text = story.author ?: "Unknown Author",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-
-                Spacer(modifier = Modifier.height(4.dp))
-
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    // Site badge
-                    story.site?.let { siteName ->
-                        Surface(
-                            color = MaterialTheme.colorScheme.primaryContainer,
-                            shape = MaterialTheme.shapes.small
-                        ) {
-                            Text(
-                                text = siteName,
-                                style = MaterialTheme.typography.labelSmall,
-                                color = MaterialTheme.colorScheme.onPrimaryContainer,
-                                modifier = Modifier.padding(horizontal = 6.dp, vertical = 2.dp)
-                            )
-                        }
-                        Spacer(modifier = Modifier.width(8.dp))
-                    }
-
-                    // Chapter count
-                    Text(
-                        text = "${story.chapters.size}/${story.totalChapters} chapters",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-
-                    Spacer(modifier = Modifier.width(8.dp))
-
-                    // Status
-                    Text(
-                        text = story.status.name,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = when (story.status) {
-                            StoryStatus.COMPLETED -> MaterialTheme.colorScheme.primary
-                            StoryStatus.ONGOING -> Color(0xFF4CAF50)
-                            StoryStatus.HIATUS -> Color(0xFFFF9800)
-                            else -> MaterialTheme.colorScheme.onSurfaceVariant
-                        }
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(8.dp))
-
-                // Tags
-                if (story.tags.isNotEmpty()) {
-                    Row {
-                        story.tags.take(3).forEach { tag ->
-                            Surface(
-                                color = MaterialTheme.colorScheme.surfaceVariant,
-                                shape = MaterialTheme.shapes.small
-                            ) {
-                                Text(
-                                    text = tag,
-                                    style = MaterialTheme.typography.labelSmall,
-                                    modifier = Modifier.padding(horizontal = 4.dp, vertical = 2.dp)
-                                )
-                            }
-                            Spacer(modifier = Modifier.width(4.dp))
-                        }
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(8.dp))
-
-                // Action buttons
-                Row {
-                    OutlinedButton(
-                        onClick = onUpdateClick,
-                        modifier = Modifier.weight(1f)
-                    ) {
-                        Icon(
-                            Icons.Default.Refresh,
-                            contentDescription = "Media image",
-                            modifier = Modifier.size(16.dp)
-                        )
-                        Spacer(modifier = Modifier.width(4.dp))
-                        Text("Update")
-                    }
-
-                    Spacer(modifier = Modifier.width(8.dp))
-
-                    Button(
-                        onClick = onDownloadClick,
-                        modifier = Modifier.weight(1f)
-                    ) {
-                        Icon(
-                            Icons.Default.Download,
-                            contentDescription = "Media image",
-                            modifier = Modifier.size(16.dp)
-                        )
-                        Spacer(modifier = Modifier.width(4.dp))
-                        Text("Download")
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun AddRedditSeriesDialog(
-    onDismiss: () -> Unit,
-    onAdd: (subreddit: String, seriesName: String, author: String) -> Unit
-) {
-    var subreddit by remember { mutableStateOf("HFY") }
-    var seriesName by remember { mutableStateOf("") }
-    var author by remember { mutableStateOf("") }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Add Reddit Series") },
-        text = {
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                OutlinedTextField(
-                    value = subreddit,
-                    onValueChange = { subreddit = it.trim() },
-                    label = { Text("Subreddit (e.g. HFY)") },
-                    modifier = Modifier.fillMaxWidth(),
-                    singleLine = true
-                )
-                OutlinedTextField(
-                    value = seriesName,
-                    onValueChange = { seriesName = it },
-                    label = { Text("Series Name (Search Query)") },
-                    modifier = Modifier.fillMaxWidth(),
-                    singleLine = true
-                )
-                OutlinedTextField(
-                    value = author,
-                    onValueChange = { author = it.trim() },
-                    label = { Text("Author (Optional)") },
-                    modifier = Modifier.fillMaxWidth(),
-                    singleLine = true
-                )
-                Text(
-                    "Use exact series title for best results.",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        },
-        confirmButton = {
-            Button(
-                onClick = { onAdd(subreddit, seriesName, author) },
-                enabled = subreddit.isNotBlank() && seriesName.isNotBlank()
-            ) {
-                Text("Download")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Cancel")
-            }
-        }
-    )
-}
-
-@Composable
-fun AddWebFictionDialog(
-    onDismiss: () -> Unit,
-    validateUrl: (String) -> Result<String>,
-    onAdd: (String) -> Unit
-) {
-    var url by remember { mutableStateOf("") }
-    var validationError by remember { mutableStateOf<String?>(null) }
-    val normalizedUrl = remember(url, validationError) {
-        if (validationError == null && url.isNotBlank()) validateUrl(url).getOrNull() else null
-    }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Text(
-                "Add Web Fiction Story",
-                fontWeight = FontWeight.Medium
-            )
-        },
-        text = {
-            Column {
-                Text(
-                    text = "Enter the URL of a story from a supported fanfiction site:",
-                    style = MaterialTheme.typography.bodyMedium,
-                    modifier = Modifier.padding(bottom = 16.dp)
-                )
-
-                OutlinedTextField(
-                    value = url,
-                    onValueChange = {
-                        url = it
-                        validationError = if (it.isBlank()) {
-                            null
-                        } else {
-                            validateUrl(it).exceptionOrNull()?.message
-                        }
-                    },
-                    label = { Text("Story URL") },
-                    placeholder = { Text("https://archiveofourown.org/works/12345") },
-                    modifier = Modifier.fillMaxWidth(),
-                    isError = validationError != null,
-                    supportingText = if (validationError != null) {
-                        { Text(validationError ?: "") }
-                    } else null,
-                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Uri)
-                )
-
-                Spacer(modifier = Modifier.height(8.dp))
-
-                normalizedUrl?.let {
-                    Text(
-                        text = "Canonical URL: $it",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                }
-
-                Text(
-                    text = "Supported sites: AO3, FFN, Royal Road, WebNovel, Wattpad, and more",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        },
-        confirmButton = {
-            Button(
-                onClick = { onAdd(normalizedUrl ?: url) },
-                enabled = url.isNotBlank() && validationError == null
-            ) {
-                Text("Add Story")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Cancel")
-            }
-        }
-    )
-}
-
-@Composable
-fun SupportedSitesDialog(
-    adultSitesEnabled: Boolean,
-    onDismiss: () -> Unit,
-    onSiteClick: (WebFictionSite) -> Unit
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Text(
-                "Supported Web Fiction Sites",
-                fontWeight = FontWeight.Medium
-            )
-        },
-        text = {
-            LazyColumn(
-                modifier = Modifier.height(400.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                items(
-                    items = WebFictionSiteType.values().filter { it != WebFictionSiteType.GENERIC },
-                    key = { siteType -> siteType.name }
-                ) { siteType ->
-                    val enabled = adultSitesEnabled || !siteType.isAdultSite()
-                    Card(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .alpha(if (enabled) 1f else 0.6f)
-                            .clickable(enabled = enabled) { onSiteClick(createWebFictionSiteFromType(siteType)) },
-                        colors = CardDefaults.cardColors(
-                            containerColor = MaterialTheme.colorScheme.surfaceVariant
-                        )
-                    ) {
-                        Row(
-                            modifier = Modifier.padding(12.dp),
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Icon(
-                                Icons.Default.Language,
-                                contentDescription = "Media image",
-                                tint = MaterialTheme.colorScheme.primary
-                            )
-                            Spacer(modifier = Modifier.width(12.dp))
-                            Column {
-                                Text(
-                                    text = getSiteDisplayName(siteType),
-                                    style = MaterialTheme.typography.titleSmall,
-                                    fontWeight = FontWeight.Medium
-                                )
-                                Text(
-                                    text = getSiteBaseUrl(siteType),
-                                    style = MaterialTheme.typography.bodySmall,
-                                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                                )
-                                if (siteType.isAdultSite()) {
-                                    Spacer(modifier = Modifier.height(6.dp))
-                                    AssistChip(
-                                        onClick = {},
-                                        enabled = false,
-                                        label = { Text("Adult Source") },
-                                        leadingIcon = {
-                                            Icon(
-                                                Icons.Default.Warning,
-                                                contentDescription = "Media image",
-                                                tint = MaterialTheme.colorScheme.tertiary
-                                            )
-                                        },
-                                        colors = AssistChipDefaults.assistChipColors(
-                                            containerColor = MaterialTheme.colorScheme.secondaryContainer,
-                                            labelColor = MaterialTheme.colorScheme.onSecondaryContainer
-                                        )
-                                    )
-                                }
-                                if (!enabled && siteType.isAdultSite()) {
-                                    Spacer(modifier = Modifier.height(6.dp))
-                                    Text(
-                                        text = "Enable in Parental Controls to view.",
-                                        style = MaterialTheme.typography.bodySmall,
-                                        color = MaterialTheme.colorScheme.error
-                                    )
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Close")
-            }
-        }
-    )
-}
 
 // Helper functions for WebFictionSiteType
 private fun getSiteDisplayName(siteType: WebFictionSiteType): String {
